requirements.txt
ConfigArgParse==1.5.3
ffmpeg-python==0.2.0
PyYAML==5.4.1
pytorch-lightning==1.8.6
pytorch-lightning-bolts==0.3.2
wandb
matplotlib==3.6.1
timm
torch==1.13.1
torchmetrics==0.11.0
torchvision==0.14.1
opencv-python
pandas==1.1.4
tqdm
